requirements.txt
astroid==2.4.2
Authlib==0.14.3
autopep8==1.5.4
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
idna==2.10
is-disposable-email==1.0.0
isort==5.6.1
isoweek==1.3.3
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongo==0.2.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pymongo==3.11.0
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
wtforms-validators==1.0.0
gunicorn==20.0.4

